Executable file
This is the executable file generated for the Patient Care Application which keeps track of patients record . Patient can register, login with thier valid credentials(username and password). The details of each customer that visits the clinic can be maintained. The underlying architecture of the design in JAVA. JAVAFX was used for beautifying and enhancing the look and feel of the application so the application is user friendly to the user.